ops -> ops,lazy
